+++ packages/snaps-utils/src/manifest/manifest.test.ts
@@ -128,6 +128,37 @@ describe('checkManifest', () => {
     expect(warnings[0]).toMatch('Missing recommended package.json properties');
   });
 
+  it('returns a warning if manifest has no defined icon', async () => {
+    const manifest = getSnapManifest();
+
+    // Remove icon
+    manifest.source.location.npm.iconPath = undefined;
+
+    await fs.writeFile(MANIFEST_PATH, JSON.stringify(manifest));
+
+    const { warnings } = await checkManifest(BASE_PATH);
+    expect(warnings).toHaveLength(1);
+    expect(warnings[0]).toMatch(
+      'No icon found in the Snap manifest. It is recommended to include an icon for the Snap. See https://docs.metamask.io/snaps/how-to/design-a-snap/#guidelines-at-a-glance for more information.',
+    );
+  });
+
+  it('returns a warning if manifest has with a non 1:1 ratio', async () => {
+    const manifest = getSnapManifest();
+
+    await fs.writeFile(
+      join(BASE_PATH, 'images/icon.svg'),
+      '<svg height="24" width="25" />',
+    );
+    await fs.writeFile(MANIFEST_PATH, JSON.stringify(manifest));
+
+    const { warnings } = await checkManifest(BASE_PATH);
+    expect(warnings).toHaveLength(1);
+    expect(warnings[0]).toMatch(
+      'The icon in the Snap manifest is not square. It is recommended to use a square icon for the Snap.',
+    );
+  });
+
   it('return errors if the manifest is invalid', async () => {
     await fs.writeFile(
       MANIFEST_PATH,