+++ packages/snaps-webpack-plugin/src/plugin.ts
@@ -84,13 +84,11 @@ export default class SnapsWebpackPlugin {
                 });
 
                 if (processed.warnings.length > 0) {
-                  compilation.warnings.push(
-                    new WebpackError(
-                      `${PLUGIN_NAME}: Bundle Warning: Processing of the Snap bundle completed with warnings.\n${processed.warnings.join(
-                        '\n',
-                      )}`,
-                    ),
+                  const webpackErrors = processed.warnings.map(
+                    (warning) => new WebpackError(warning),
                   );
+
+                  compilation.warnings.push(...webpackErrors);
                 }
 
                 const replacement = processed.sourceMap
@@ -159,11 +157,7 @@ export default class SnapsWebpackPlugin {
 
         if (warnings.length > 0) {
           compilation.warnings.push(
-            new WebpackError(
-              `${PLUGIN_NAME}: Manifest Warning: Validation of snap.manifest.json completed with warnings.\n${warnings.join(
-                '\n',
-              )}`,
-            ),
+            ...warnings.map((warning) => new WebpackError(warning)),
           );
         }
       }