+++ packages/examples/packages/dialogs/CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0]
+### Added
+- Add a link to confirmation dialog ([#2112](https://github.com/MetaMask/snaps/pull/2112))
+
+### Changed
+- Remove snap icon ([#2189](https://github.com/MetaMask/snaps/pull/2189))
+
 ## [2.1.0]
 ### Changed
 - Use `@metamask/snaps-sdk` package ([#1930](https://github.com/MetaMask/snaps/pull/1930),
@@ -44,7 +51,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - The version of the package no longer needs to match the version of all other
     MetaMask Snaps packages.
 
-[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/dialog-example-snap@2.1.0...HEAD
+[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/dialog-example-snap@2.2.0...HEAD
+[2.2.0]: https://github.com/MetaMask/snaps/compare/@metamask/dialog-example-snap@2.1.0...@metamask/dialog-example-snap@2.2.0
 [2.1.0]: https://github.com/MetaMask/snaps/compare/@metamask/dialog-example-snap@2.0.1...@metamask/dialog-example-snap@2.1.0
 [2.0.1]: https://github.com/MetaMask/snaps/compare/@metamask/dialog-example-snap@2.0.0...@metamask/dialog-example-snap@2.0.1
 [2.0.0]: https://github.com/MetaMask/snaps/compare/@metamask/dialog-example-snap@1.0.0...@metamask/dialog-example-snap@2.0.0