+++ packages/examples/packages/get-file/CHANGELOG.md
@@ -6,11 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1]
+### Changed
+- Remove snap icon ([#2189](https://github.com/MetaMask/snaps/pull/2189))
+
 ## [1.1.0]
 ### Changed
 - Use `@metamask/snaps-sdk` package ([#1946](https://github.com/MetaMask/snaps/pull/1946))
-    - This package replaces the `@metamask/snaps-types` and
-    - `@metamask/snaps-ui` packages, and is much more lightweight.
+  - This package replaces the `@metamask/snaps-types` and
+  - `@metamask/snaps-ui` packages, and is much more lightweight.
 
 ## [1.0.1]
 ### Fixed
@@ -20,7 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `snap_getFile` example Snap ([#1836](https://github.com/MetaMask/snaps/pull/1836), [#1858](https://github.com/MetaMask/snaps/pull/1858))
 
-[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/get-file-example-snap@1.1.0...HEAD
+[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/get-file-example-snap@1.1.1...HEAD
+[1.1.1]: https://github.com/MetaMask/snaps/compare/@metamask/get-file-example-snap@1.1.0...@metamask/get-file-example-snap@1.1.1
 [1.1.0]: https://github.com/MetaMask/snaps/compare/@metamask/get-file-example-snap@1.0.1...@metamask/get-file-example-snap@1.1.0
 [1.0.1]: https://github.com/MetaMask/snaps/compare/@metamask/get-file-example-snap@1.0.0...@metamask/get-file-example-snap@1.0.1
 [1.0.0]: https://github.com/MetaMask/snaps/releases/tag/@metamask/get-file-example-snap@1.0.0