+++ packages/examples/packages/rollup-plugin/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.1]
+### Changed
+- Remove snap icon ([#2189](https://github.com/MetaMask/snaps/pull/2189))
+
 ## [2.1.0]
 ### Changed
 - Use `@metamask/snaps-sdk` package ([#1946](https://github.com/MetaMask/snaps/pull/1946))
@@ -34,7 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - The version of the package no longer needs to match the version of all other
     MetaMask Snaps packages.
 
-[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/rollup-plugin-example-snap@2.1.0...HEAD
+[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/rollup-plugin-example-snap@2.1.1...HEAD
+[2.1.1]: https://github.com/MetaMask/snaps/compare/@metamask/rollup-plugin-example-snap@2.1.0...@metamask/rollup-plugin-example-snap@2.1.1
 [2.1.0]: https://github.com/MetaMask/snaps/compare/@metamask/rollup-plugin-example-snap@2.0.1...@metamask/rollup-plugin-example-snap@2.1.0
 [2.0.1]: https://github.com/MetaMask/snaps/compare/@metamask/rollup-plugin-example-snap@2.0.0...@metamask/rollup-plugin-example-snap@2.0.1
 [2.0.0]: https://github.com/MetaMask/snaps/compare/@metamask/rollup-plugin-example-snap@1.0.0...@metamask/rollup-plugin-example-snap@2.0.0